+++ src/core/evaluate.py
@@ -6,7 +6,7 @@ from torchvision import transforms
 from PIL import Image
 import os
 import yaml
-import models as models
+import src.core.models as models
 import dataset_preparation
 import matplotlib.pyplot as plt
 from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, precision_score, recall_score, f1_score, roc_auc_score, roc_curve
