+++ src/core/train.py
@@ -5,8 +5,8 @@ import torch.optim as optim
 from torch.utils.data import DataLoader, random_split
 from torchvision import transforms
 from config import get_model_config
-from evaluate import evaluate_model, generate_confusion_matrix
-import models as models
+from src.core.evaluate import evaluate_model, generate_confusion_matrix
+import src.core.models as models
 import dataset_preparation
 from tqdm import tqdm
 import yaml
